backend-flask-api/requirements.txt
# --- Core Server ---
Flask>=2.2,<3.0
Flask-Cors>=4.0.0
gunicorn==22.0.0
requests>=2.31.0

# --- Google AI & Cloud ---
google-genai>=0.1.0
vertexai>=1.71.1
google-cloud-aiplatform==1.71.1
# Google Ads client with support for newer API versions (v22+)
google-ads

# --- Compatibility Fixes (protobuf / grpc) ---
protobuf>=3.20.2,<6.0.0dev
grpcio-status==1.71.2

# --- Utilities & Image Processing ---
opencv-python>=4.9.0.80
PyYAML>=6.0.0

# --- Dev Convenience ---
python-dotenv>=1.0.1
